mips: XIP kernel support
This adds XIP support for MIPS that allows the kernel to execute from ROM.  The
data and bss sections are copied to RAM so the linker script is heavily modified
to support this.  In addition, there are runtime generated functions that had to
be moved outside of readonly sections.

This is highly experimental and is not likely to work in all cases.