pic32mz: Convert platform setup code to device tree.
DT was originally not used because it adds significant space overhead for the
XIP kernel. However, traditional platform setup was broken for many device
drivers in previous commits and those same commits depend on having DT support.

With zboot, the space issue is less of a conscern because it provides an easy
alternative.  In any event, maintaining DT config and traditional platform setup
is just messy overhead so this is a start at converting it all to DT.