pic32: Remove PIC32 conditional kernel_entry_setup out of arch/mips code.
This moves the condition on whether to call kernel_entry_setup out of the
arch/mips code since the implementation is already PIC32 specific. In the PIC32
implementation the condition is then used to only not initialize cache if ZBOOT
is used because that already handles cache setup like a bootloader would.

In the future, this whole PIC32 implementation will likely need to be removed
in favor of the bootloader intitialzing cache.