pic32-ocmp: Fix PWM glitch when changing duty cycle of enabled module.
This fixes a glitch on all PWMs when changing only the PWM duty cycle of a
single PWM.

pic32_pwm_config() is a generic function used to configure all parameters of the
PWM and therefore it always sets the mode, even when the module is enabled.
This change indirectly prevents writing to the register if we are about to write
the same value anyway.

The hint that led toward this fix from the FRM is as follows:

The Output Compare module must be turned OFF by the user application (i.e.,
clear the OCM<2:0> bits (OCxCON<2:0>)) before switching to a new mode.
Changing modes while the module is in operation may produce unexpected results.